+++ CHANGELOG.md
@@ -22,4 +22,11 @@
 - Add auto-redirect feature
 - Add Pageviews feature
 ## 1.1
-- Remove credentials,keepalive and header content type from header 
+- Remove credentials, keepalive and header content type from header 
+## 1.2.0
+- Improvement: Use navigator.sendBeacon instead of fetch when supported by the browser for sending data in the background.
+- Improvement: Add slug translation support for better handling of translated URLs.
+- Improvement: Add a new “dynamics” option in the admin interface to give more control over dynamic content behavior.
+- Improvement: Add reverse translation support for Craft’s search base to improve multilingual search handling.
+## 1.2.1
+- Improvement: merge all dynamic selectors into the whitelist before injecting Weglot scripts in the DynamicsService.